grocer/grocer/agent/requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiosignal==1.3.1
alembic==1.13.3
aniso8601==9.0.1
annotated-types==0.7.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.1.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
black==22.6.0
bleach==4.1.0
blinker==1.8.2
boto3==1.24.28
botocore==1.27.96
build==1.2.2
cachetools==5.3.2
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.4
chroma-hnswlib==0.7.6
chromadb==0.5.11
click==8.1.7
cloudpickle==3.0.0
coloredlogs==15.0.1
comm==0.1.2
contourpy==1.0.5
cryptography==39.0.1
cycler==0.11.0
Cython==0.29.32
databricks-connect==14.3.2
databricks-sdk==0.20.0
databricks-sql-connector==3.4.0
databricks-vectorsearch==0.40
dataclasses-json==0.6.7
dbus-python==1.2.18
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
dirtyjson==1.0.8
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
docker==7.1.0
docopt==0.6.2
docstring-to-markdown==0.11
durationpy==0.7
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
facets-overview==1.1.1
fastapi==0.115.0
fastjsonschema==2.19.1
filelock==3.14.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.25.0
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.28.1
googleapis-common-protos==1.62.0
graphene==3.3
graphql-core==3.2.4
graphql-relay==3.2.0
greenlet==3.1.1
grpcio==1.62.0
grpcio-status==1.62.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.25.1
humanfriendly==10.0
idna==3.4
importlib_metadata==8.4.0
importlib_resources==6.4.5
ipyflow-core==0.0.198
ipykernel==6.25.0
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==7.7.2
itsdangerous==2.2.0
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.1.2
jiter==0.5.0
jmespath==0.10.0
joblib==1.2.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.17.3
jupyter-client==7.3.4
jupyter-server==1.23.4
jupyter_core==5.2.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keyring==23.5.0
kiwisolver==1.4.4
kubernetes==31.0.0
langchain==0.3.1
langchain-community==0.3.1
langchain-core==0.3.6
langchain-experimental==0.3.2
langchain-openai==0.2.1
langchain-text-splitters==0.3.0
langchainhub==0.1.21
langsmith==0.1.129
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
llama-cloud==0.1.0
llama-index==0.11.14
llama-index-agent-openai==0.3.4
llama-index-cli==0.3.1
llama-index-core==0.11.14
llama-index-embeddings-huggingface==0.3.1
llama-index-embeddings-openai==0.2.5
llama-index-indices-managed-llama-cloud==0.4.0
llama-index-legacy==0.9.48.post3
llama-index-llms-databricks==0.2.0
llama-index-llms-openai==0.2.9
llama-index-llms-openai-like==0.2.0
llama-index-multi-modal-llms-openai==0.2.1
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.2
llama-index-readers-llama-parse==0.3.0
llama-index-readers-wikipedia==0.2.0
llama-index-vector-stores-chroma==0.2.0
llama-index-vector-stores-faiss==0.2.1
llama-parse==0.5.6
lxml==4.9.1
lz4==4.3.3
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.1
marshmallow==3.22.0
matplotlib==3.7.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
minijinja==2.2.0
mistune==3.0.2
mlflow==2.16.2
mlflow-skinny==2.16.2
mmh3==5.0.1
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==0.4.3
nbclassic==0.5.2
nbclient==0.5.13
nbconvert==7.16.4
nbformat==5.7.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
nodeenv==1.8.0
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.19.2
openai==1.50.2
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.7
overrides==7.7.0
packaging==23.2
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pipreqs==0.5.0
platformdirs==2.5.2
plotly==5.9.0
pluggy==1.0.0
posthog==3.6.6
prometheus-client==0.14.1
prompt-toolkit==3.0.36
protobuf==4.25.3
psutil==5.9.0
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==16.1.0
pyarrow-hotfix==0.5
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyccolo==0.0.52
pycparser==2.21
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pyflakes==3.1.0
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pyodbc==4.0.32
pyparsing==3.0.9
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyright==1.1.294
pyrsistent==0.18.0
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-dotenv==1.0.1
python-lsp-jsonrpc==1.1.1
python-lsp-server==1.8.0
pytoolconfig==1.2.5
pytz==2022.7
PyYAML==6.0.2
pyzmq==23.2.0
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.1
rope==1.7.0
rsa==4.9
s3transfer==0.6.2
safetensors==0.4.5
scikit-learn==1.1.1
scipy==1.10.0
seaborn==0.12.2
SecretStorage==3.3.1
Send2Trash==1.8.0
sentence-transformers==3.1.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.2.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.35
sqlparse==0.5.1
ssh-import-id==5.11
stack-data==0.2.0
starlette==0.38.6
statsmodels==0.13.5
striprtf==0.0.26
sympy==1.13.3
tenacity==8.5.0
terminado==0.17.1
threadpoolctl==2.2.0
thrift==0.20.0
tiktoken==0.7.0
tinycss2==1.2.1
tokenize-rt==4.2.1
tokenizers==0.20.0
tomli==2.0.1
torch==2.4.1
tornado==6.1
tqdm==4.66.5
traitlets==5.7.1
transformers==4.45.1
triton==3.0.0
typer==0.12.5
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing_extensions==4.12.2
ujson==5.4.0
unattended-upgrades==0.1
urllib3==1.26.14
uvicorn==0.31.0
uvloop==0.20.0
virtualenv==20.16.7
wadllib==1.3.6
watchfiles==0.24.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
websockets==13.1
Werkzeug==3.0.4
whatthepatch==1.0.2
widgetsnbextension==3.6.1
wikipedia==1.4.0
wrapt==1.16.0
yapf==0.33.0
yarg==0.1.9
yarl==1.13.1
zipp==1.0.0
